Preserve generic type overrides in Sorbet casts
