+++ lib/tapioca/internal.rb
@@ -12,6 +12,7 @@ require "tapioca/runtime/dynamic_mixin_compiler"
 require "tapioca/sorbet_ext/backcompat_patches"
 require "tapioca/sorbet_ext/name_patch"
 require "tapioca/sorbet_ext/generic_name_patch"
+require "tapioca/sorbet_ext/generic_type_patch"
 require "tapioca/sorbet_ext/proc_bind_patch"
 require "tapioca/sorbet_ext/void_patch"
 require "tapioca/runtime/generic_type_registry"
