+++ lib/tapioca/sorbet_ext/generic_name_patch.rb
@@ -82,26 +82,6 @@ module T
       prepend GenericPatch
     end
   end
-
-  module Utils
-    module Private
-      module PrivateCoercePatch
-        def coerce_and_check_module_types(val, check_val, check_module_type)
-          if val.is_a?(Tapioca::TypeVariableModule)
-            val.coerce_to_type_variable
-          elsif val.respond_to?(:__tapioca_override_type)
-            val.__tapioca_override_type
-          else
-            super
-          end
-        end
-      end
-
-      class << self
-        prepend(PrivateCoercePatch)
-      end
-    end
-  end
 end
 
 module Tapioca
